+++ iqscan/__main__.py
@@ -1,0 +1,4 @@
+import sys
+from .iq_scan import main
+
+sys.exit(main())